requirements.txt
cachetools==5.5.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==46.0.2
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.31.0
idna==3.10
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.1.4
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
